requirements.txt
# -------------------------------
# Core MCP + Runtime
# -------------------------------
fastmcp>=0.4.0
python-dotenv>=1.0.0
loguru>=0.7.0

# -------------------------------
# Web stack
# -------------------------------
fastapi>=0.110.0
starlette>=0.36.3
uvicorn>=0.27.0
pydantic>=2.6
pydantic-settings>=2.2.1

# -------------------------------
# Database (SQLite)
# -------------------------------
aiosqlite>=0.19.0

# -------------------------------
# SQL stack (optional / kept intentionally)
# -------------------------------
databases>=0.7.0
sqlalchemy>=2.0.0
alembic>=1.10.0
